requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-bootstrap5==0.1.dev1
Flask-Login==0.6.3
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg==3.2.3
psycopg2==2.9.10
python-dotenv==1.0.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
WTForms==3.2.1